% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{add_areasqkm}
\alias{add_areasqkm}
\title{Compute km2 area
Short hand for safely computing area in sqkm and returning as numeric vector.}
\usage{
add_areasqkm(x)
}
\arguments{
\item{x}{sf object}
}
\value{
numeric vector
}
\description{
Compute km2 area
Short hand for safely computing area in sqkm and returning as numeric vector.
}
\examples{
library(sf)
nc = st_read(system.file("shape/nc.shp", package="sf"))
add_areasqkm(nc[1,])
}
