% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_area_sqkm_to_crosswalk.R
\name{add_areasqkm_to_crosswalk}
\alias{add_areasqkm_to_crosswalk}
\title{Add small area to crosswalk}
\usage{
add_areasqkm_to_crosswalk(crosswalk, comid = "hf_id")
}
\arguments{
\item{crosswalk}{an existing crosswalk table}

\item{comid}{the shared ID}
}
\value{
data.frame
}
\description{
Add small area to crosswalk
}
