% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_handlers.R
\name{add_flowpath_edge_list}
\alias{add_flowpath_edge_list}
\title{Generate Catchment Network Table}
\usage{
add_flowpath_edge_list(gpkg)
}
\arguments{
\item{gpkg}{a  hydrofabric gpkg}
}
\value{
data.frame with ID, toID, length, area, and levelpath
}
\description{
Generate Catchment Network Table
}
