% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{add_hydroseq}
\alias{add_hydroseq}
\title{Add hydrosequence}
\usage{
add_hydroseq(flowpaths)
}
\arguments{
\item{flowpaths}{sf object (LINESTRING)}
}
\value{
sf object
}
\description{
Add hydrosequence
}
