% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{add_lengthkm}
\alias{add_lengthkm}
\title{Compute length in kilometers}
\usage{
add_lengthkm(x)
}
\arguments{
\item{x}{LINESTRING sf object}
}
\value{
numeric vector
}
\description{
Compute length in kilometers
}
