% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{add_lengthmap}
\alias{add_lengthmap}
\title{Add Length Map to Refactored Network}
\usage{
add_lengthmap(flowpaths)
}
\arguments{
\item{flowpaths}{a refactored flowpath network containing an member_COMID column}
}
\value{
sf object
}
\description{
This function replicated the member_COMID column of a refactored 
network but adds a new notion. Following each COMID is '.' which is proceeded 
by the fraction of that COMID making up the new flowpath. For example 101.1 
would indicate 100% of COMID 101 is in the new ID. 
Equally 101.05 would indicate 50% of COMID 101 is present in the new ID
}
\examples{
path <- system.file("extdata/walker_reconcile.gpkg", package = "hyRefactor")
fps  = read_sf(path) \%>\% 
  add_lengthmap()
}
