% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_handlers.R
\name{add_lookup_table}
\alias{add_lookup_table}
\title{Generate Lookup table for refactored or aggregated network}
\usage{
add_lookup_table(
  gpkg = NULL,
  refactored_gpkg = NULL,
  reconciled_layer = "flowpaths"
)
}
\arguments{
\item{gpkg}{character path to gpkg containing aggregated network. Omit for 
refactored network lookup table creation.}

\item{refactored_gpkg}{character path to the gpkg for the refactored network 
used to create the aggregate network. If no aggregatedd gpkg is passed in, a 
lookup table will be added to this gpkg.}

\item{reconciled_layer}{character path layer name containing fully 
reconciled flowpaths. Ignored for aggregated network lookup table creation.}
}
\value{
file path to modified gpkg
}
\description{
Generate Lookup table for refactored or aggregated network
}
