% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_handlers.R
\name{add_mapped_pois}
\alias{add_mapped_pois}
\title{Add a mapped_POI layer to network_list}
\usage{
add_mapped_pois(
  network_list,
  type = c("HUC12", "Gages", "TE", "NID", "WBIn", "WBOut"),
  refactored_gpkg = NULL,
  verbose = TRUE
)
}
\arguments{
\item{network_list}{a list with flowpath and catchment data}

\item{refactored_gpkg}{a (optional) path to}

\item{verbose}{should messages be emited?}
}
\value{
list()
}
\description{
Add a mapped_POI layer to network_list
}
