% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{add_measures}
\alias{add_measures}
\title{Add/sync/update length and area measures}
\usage{
add_measures(flowpaths, divides)
}
\arguments{
\item{cat}{sf object (POLYGON)}
}
\value{
list
}
\description{
Add/sync/update length and area measures
}
