% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{add_prefix}
\alias{add_prefix}
\title{Add Prefixes to Topological Data}
\usage{
add_prefix(
  topo,
  hf_prefix = "cat-",
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-",
  coastal_nexus_prefix = "cnx-",
  internal_nexus_prefix = "inx-"
)
}
\arguments{
\item{topo}{A data frame containing topological data. Must include columns `topo_type`, `type`, `id`, and `toid`.}

\item{hf_prefix}{A character string to prefix `id` values in rows where `topo_type == "network"`. Default is `"cat-"`.}

\item{nexus_prefix}{A character string to prefix `toid` values in rows where `type` does not match specific cases (`terminal`, `coastal`, or `internal`). Default is `"nex-"`.}

\item{terminal_nexus_prefix}{A character string to prefix `toid` values where `type == "terminal"`. Default is `"tnx-"`.}

\item{coastal_nexus_prefix}{A character string to prefix `toid` values where `type == "coastal"`. Default is `"cnx-"`.}

\item{internal_nexus_prefix}{A character string to prefix `toid` values where `type == "internal"`. Default is `"inx-"`.}
}
\value{
A data frame with updated `id` and `toid` values, including appropriate prefixes based on the type of topology and the `type` column. The resulting data frame contains the columns `id`, `toid`, `type`, and any additional columns with names containing `"vpu"`.
}
\description{
This function adds specified prefixes to the `id` and `toid` columns of a topological data frame based on the type of topology and its context (e.g., network or nexus).
}
