% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{agg_length_area}
\alias{agg_length_area}
\title{Enforces area and length grouping}
\usage{
agg_length_area(l, a, lthres, athres)
}
\arguments{
\item{l}{a vector of lengths}

\item{a}{a vector of areas}

\item{lthres}{a minimum length that must be achieved}

\item{athres}{a minimum length that must be achieved}
}
\value{
a vector of length(a) containing grouping indexes
}
\description{
This function takes a vector of area's and length's and returns a
grouping vector that enforces the grouping of lengths and areas less then defined thresholds
}
