% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{aggregate_along_mainstems}
\alias{aggregate_along_mainstems}
\title{Aggregate along network mainstems}
\usage{
aggregate_along_mainstems(
  network_list,
  ideal_size_sqkm,
  min_area_sqkm,
  min_length_km,
  verbose = TRUE,
  cache_file = NULL
)
}
\arguments{
\item{network_list}{a list containing flowline and catchment `sf` objects}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}

\item{ideal_size}{The ideal size of output hydrofabric catchments}

\item{term_cut}{cutoff integer to define terminal IDs}
}
\value{
a list containing aggregated and validated flowline and catchment `sf` objects
}
\description{
Given a set of ideal catchment sizes, plus the
minimum allowable catchment size and segment length, aggregate the network along mainstems.
}
