% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{aggregate_sets}
\alias{aggregate_sets}
\title{Aggregate Sets by Index Table}
\usage{
aggregate_sets(network_list, index_table)
}
\arguments{
\item{network_list}{a list of flowpaths and catchments}

\item{index_table}{index table to aggregate with}
}
\value{
a list of catchments and flowpaths that have been validated
}
\description{
Aggregate Sets by Index Table
}
