% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_distribution.R
\name{aggregate_to_distribution}
\alias{aggregate_to_distribution}
\title{Aggregate Network to Uniform Size}
\usage{
aggregate_to_distribution(
  gpkg = NULL,
  flowpath = NULL,
  divide = NULL,
  outlets = NULL,
  ideal_size_sqkm = 10,
  min_length_km = 1,
  min_area_sqkm = 3,
  outfile = NULL,
  log = TRUE,
  overwrite = FALSE,
  cache = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gpkg}{a path to a gpkg}

\item{flowpath}{If gpkg is NULL, then an sf data.frame, otherwise a the layer name. See details.}

\item{divide}{If gpkg is NULL, then an sf data.frame, otherwise a the layer name. See details.}

\item{outlets}{data.frame with mandatory "ID" column and optional "POI_ID" column. "ID" must be identifiers from
flowpath and divide data.frames and POI ID must be unique.}

\item{ideal_size_sqkm}{The ideal size of catchments (default = 10 sqkm)}

\item{min_length_km}{The minimum allowable length of flowpath features (default = 1 km)}

\item{min_area_sqkm}{The minimum allowable area of catchment features (default = 3 sqkm)}

\item{outfile}{of not NULL, where to write the output files}

\item{log}{a filepath to write messages to or Boolean (TRUE = print to console; FALSE = no messages)}

\item{overwrite}{overwrite existing gf file. Default is FALSE}

\item{verbose}{print status updates. Default = TRUE}

\item{nexus_locations}{a data.frame with columns specifying the ID, and the nexus type.}
}
\value{
if outfile = TRUE, a file path, else a list object
}
\description{
This function aggregates a network to a desired size distribution while
enforcing minimum flowpath legnths and catchment areas. Additionally a set of explicit nexus
locations can be provided over which the network cannot be aggregated (see poi_to_outlet)
}
\details{
If gpkg is not NULL, divide and flowpath can be left NULL as well. The code attempts to
infer the correct layers. The divides layer will be the one including the word "divide" or "catchment" and the
flowpath layer will be the one including 'flowpath' or 'flowline'. If no layers, or more then one layer are deemed possible
for each input, then the function will stop and ask for explicit names.
}
