% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_style.R
\name{append_style}
\alias{append_style}
\title{Append a hydrofabric style to a hydrofabric GeoPackage}
\usage{
append_style(gpkg_path, layer_names)
}
\arguments{
\item{gpkg_path}{Path to GeoPackage}

\item{layer_names}{character vector of names to append styles for.
These names must be in the package QML directory.}
}
\description{
Append a hydrofabric style to a hydrofabric GeoPackage
}
