% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo.R
\name{apply_nexus_topology}
\alias{apply_nexus_topology}
\title{Apply Nexus Topology}
\usage{
apply_nexus_topology(
  gpkg,
  catchments = NULL,
  flowpaths = NULL,
  vpu = NA,
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-",
  coastal_nexus_prefix = "cnx-",
  internal_nexus_prefix = "inx-",
  catchment_prefix = "cat-",
  waterbody_prefix = "wb-",
  term_add = 1e+09,
  term_filter = NULL,
  verbose = TRUE,
  enforce_dm = FALSE,
  export_gpkg = NULL
)
}
\arguments{
\item{nexus_prefix}{character prefix for nexus IDs}

\item{terminal_nexus_prefix}{character prefix for terminal nexus IDs}

\item{coastal_nexus_prefix}{character prefix for coastal nexus IDs}

\item{internal_nexus_prefix}{character prefix for internal nexus IDs}

\item{catchment_prefix}{character prefix for catchment IDs}

\item{waterbody_prefix}{character prefix for catchment IDs}

\item{enforce_dm}{should the data model be validated prior to writing?}

\item{export_gpkg}{file path to write new data. If NULL list object is returned}

\item{network_list}{list containing flowpath and catchment `sf` objects}
}
\value{
list or file path
}
\description{
This function enforces the nexus-->flowpath topology and adds nexus locations,
a catchment edge list, a flowpath edge list, and a lookup_table to the
network_list object.
}
