% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{assign_global_identifiers}
\alias{assign_global_identifiers}
\title{Update Hydrofabric Identifiers
For a given set of hydrofabric geopackages, update the ID and toID values to 
be globally unique.}
\usage{
assign_global_identifiers(
  gpkgs = NULL,
  outfiles = NULL,
  flowpath_layer = "flowpaths",
  divide_layer = "divides",
  network_layer = "network",
  overwrite = FALSE,
  term_add = 1e+09,
  modifications = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gpkgs}{a vecotor of file.paths to that define the global network}

\item{outfiles}{a vector of file.paths to write to}

\item{flowpath_layer}{the layer name containing flowpaths}

\item{divide_layer}{the layer name containing divides}

\item{network_layer}{the name of layer containing the hydrologic network}

\item{overwrite}{overwrite existing files?}

\item{term_add}{value to be added to all terminal IDs}

\item{verbose}{emit messages}
}
\value{
a data.frame
}
\description{
Update Hydrofabric Identifiers
For a given set of hydrofabric geopackages, update the ID and toID values to 
be globally unique.
}
