% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{assign_global_terminal_identifiers}
\alias{assign_global_terminal_identifiers}
\title{Update Hydrofabric Terminal Identifiers
For a given set of hydrofabric geopackages, update terminal toID values to 
be gloablaly unique.}
\usage{
assign_global_terminal_identifiers(
  meta,
  flowpath_layer = "flowpaths",
  divide_layer = "divides",
  lookup_table_layer = "lookup_table",
  flowpath_edge_list = "flowpath_edge_list",
  verbose = TRUE,
  term_add = 1e+09
)
}
\arguments{
\item{flowpath_layer}{the layer name containing flowpaths}

\item{lookup_table_layer}{the layer name containing flowpaths}

\item{verbose}{emit messages}

\item{term_add}{a numeric value to add to the terminal ids to facilitate 
easy discovery, and to avoid conflict with know IDs}

\item{gpkgs}{a metadata data.frame resulting from `assign_global_identifiers`}

\item{catchment_network_layer}{the layer name containing flowpaths}
}
\value{
data.frame
}
\description{
Update Hydrofabric Terminal Identifiers
For a given set of hydrofabric geopackages, update terminal toID values to 
be gloablaly unique.
}
