% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{assign_id}
\alias{assign_id}
\title{Index a Vector by Cumulative Sum}
\usage{
assign_id(x, athres)
}
\arguments{
\item{x}{a vector of values}

\item{athres}{the ideal cumulative size of each group
Cumulative sums will get as close to this value without exceeding it}
}
\value{
a vector of length(a)
}
\description{
Index a Vector by Cumulative Sum
}
