% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_inward.R
\name{build_collapse_table}
\alias{build_collapse_table}
\title{Build Headwater Collapse Table}
\usage{
build_collapse_table(network_list, min_area_sqkm = 3, min_length_km = 1)
}
\arguments{
\item{network_list}{a list containing flowpath and catchment `sf` objects}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}
}
\value{
a 2 column data.frame with {id, becomes}
}
\description{
Identifies small (pathlength or area) headwater catchments and returns a data.frame
with the current ID and the feature ID it should collapse into (becomes).
Headwaters are segments in which there are no inflows (!ID %in% toID).
}
