% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{build_new_id_table}
\alias{build_new_id_table}
\title{Build a new ID table}
\usage{
build_new_id_table(
  meta,
  index,
  network_layer = "network",
  network_lookup = "network_lookup",
  term_add = 1e+09,
  modifications = NULL
)
}
\arguments{
\item{network_layer}{the name of layer containing the hydrologic network}

\item{gpkgs}{a row of network metadata built with `network_metadata`}

\item{flowpath_layer}{the layer name containing flowpaths}

\item{divide_layer}{the layer name containing divides}
}
\value{
data.frame
}
\description{
Build a new ID table
}
