% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{clean_geometry}
\alias{clean_geometry}
\title{Clean Catchment Geometry}
\usage{
clean_geometry(
  catchments,
  flowlines = NULL,
  fl_ID = NULL,
  ID = "ID",
  keep = NULL,
  crs = 5070,
  grid = 9e-04,
  gb = 8,
  force = FALSE,
  sys = NULL
)
}
\arguments{
\item{catchments}{catchments geometries to fix}

\item{flowlines}{flowlines geometries to filter largest unit (optional)}

\item{fl_ID}{flowlines unique identifier}

\item{ID}{name of uniquely identifying column}

\item{keep}{proportion of points to retain in geometry simplification
(0-1; default 0.05). See \code{\link[rmapshaper]{ms_simplify}}.
If NULL, then no simplification will be executed.}

\item{crs}{integer or object compatible with sf::st_crs coordinate reference.
Should be a projection that supports area-calculations.}

\item{gb}{The amount of heap memory to be allocated when force = TRUE}

\item{force}{should the mapshaper/mapshaper-xl binaries be used directly for simplification?}

\item{sys}{logical should the mapshaper system library be used. If NULL 
the system library will be used if available.}
}
\value{
sf object
}
\description{
Fixes geometry issues present in catchments derived from DEMs.
These include, but are not limited to disjoint polygon fragments, artifacts
from the DEM used to generate the catchments, and non-valid geometry topologies.
A secondary goal of this functions is to provide a way to reduce the data column
of the catchments by offering a topology preserving simplification
through \code{\link[rmapshaper]{ms_simplify}}.
Generally a "keep" parameter of .9 seems appropriate for the resolution of
the data but can be modified in function
}
