% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_style.R
\name{create_style_row}
\alias{create_style_row}
\title{Create a style row from the given parameters.}
\usage{
create_style_row(gpkg_path, layer_name, style_name, style_qml)
}
\arguments{
\item{gpkg_path}{Path to GeoPackage}

\item{layer_name}{Layer to create style for}

\item{style_name}{Name of the new style}

\item{style_qml}{QML contents for the new style}
}
\value{
1-row data.frame in the layer_styles table schema
}
\description{
Create a style row from the given parameters.
}
\keyword{internal}