% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{cs_group}
\alias{cs_group}
\title{Cumulative sum area grouping}
\usage{
cs_group(areas, lengths, exclude, ideal_size_sqkm, amin, lmin)
}
\arguments{
\item{areas}{a vector of areas}

\item{lengths}{a vector of lengths}

\item{exclude}{a vector of equal length to areas and lengths. Any non NA value will be used to enforce an aggregation break}

\item{ideal_size_sqkm}{a vector of areas}

\item{amin}{a threshold, or target, cumulative size}

\item{lmin}{a threshold, or target, cumulative size}
}
\value{
a vector of length(areas) containing grouping indexes
}
\description{
This function takes a vector of areas and lengths and returns a
index vector that combines them towards an ideal aggregate area (ideal_size_sqkm). While enforcing a minimum area (amin) and length (lmin).
Additionally, this function can take a set of indexes to exclude over which the network cannot be aggregated.
}
