% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_inward.R
\name{define_touch_id}
\alias{define_touch_id}
\title{Identify intersection types and downstream topology}
\usage{
define_touch_id(flowpaths, term_cut = 1e+09)
}
\arguments{
\item{flowpaths}{sf LINESTRING}
}
\value{
data.frame with id, type, touches, touches_toID columns
}
\description{
Identify intersection types and downstream topology
}
