% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fdr_fac.R
\name{download_elev}
\alias{download_elev}
\title{Download Elevation and Derivatives}
\usage{
download_elev(product, out_dir, regions = NULL)
}
\arguments{
\item{product}{character DEM, hydroDEM, or FDRFAC.}

\item{out_dir}{path to directory to store output.}

\item{regions}{character vector of two digit hydrologic}
}
\description{
Download Elevation and Derivatives
}
