% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fdr_fac.R
\name{download_fdr_fac}
\alias{download_fdr_fac}
\title{Download FDR FAC}
\usage{
download_fdr_fac(out_dir, regions = NULL)
}
\arguments{
\item{out_dir}{path to directory to store output.}

\item{regions}{character vector of two digit hydrologic 
regions to download. Omit for all.}
}
\description{
Download FDR FAC
}
