% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{drop_extra_features}
\alias{drop_extra_features}
\title{Remove non-coincident Network Features
Remove non-coincident flowlines and catchment pairs from a network list}
\usage{
drop_extra_features(network_list, verbose)
}
\arguments{
\item{network_list}{a list containing flowpaths and catchments}

\item{verbose}{should message be emitted?}
}
\value{
a list containing flowpaths and catchments
}
\description{
Remove non-coincident Network Features
Remove non-coincident flowlines and catchment pairs from a network list
}
