% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_subsetting.R
\name{find_origin}
\alias{find_origin}
\title{Find ID from location}
\usage{
find_origin(gpkg, pt, catchment_name = "aggregate_divides")
}
\arguments{
\item{gpkg}{path to a hydrofabric}

\item{pt}{a spatial point (sf)}
}
\value{
a waterbody ID (character)
}
\description{
Find ID from location
}
