% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R, R/topo.R
\name{flush_prefix}
\alias{flush_prefix}
\title{Flush existing ID prefixes
Given a data object and column, remove a prefix and adjoining "-"}
\usage{
flush_prefix(input, col)

flush_prefix(input, col)
}
\arguments{
\item{input}{A data frame containing the columns to be processed.}

\item{col}{A character vector specifying the names of the columns from which to remove prefixes.}
}
\value{
data object with updated column

The input data frame with the specified columns updated. The values in these columns are converted to numeric, retaining only the portion after the last `-` character.
}
\description{
This function removes prefixes from specified columns in a data frame by extracting the numeric portion of the values after the last `-` character.
}
\details{
- The function processes each specified column by removing the prefix up to and including the last `-` character using a regular expression.
- The updated columns are converted to numeric values.
}
