% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_handlers.R
\name{generate_catchment_network}
\alias{generate_catchment_network}
\title{Generate Catchment Network Table}
\usage{
generate_catchment_network(gpkg)
}
\arguments{
\item{gpkg}{a  hydrofabric gpkg}
}
\value{
data.frame with ID, toID, length, area, and levelpath
}
\description{
Generate Catchment Network Table
}
