% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi_handlers.R
\name{generate_lookup_table}
\alias{generate_lookup_table}
\title{Generate Lookup table for Aggregated Network}
\usage{
generate_lookup_table(gpkg = NULL, refactored_gpkg = NULL)
}
\arguments{
\item{gpkg}{to current network. A look up table will be added here as a layer called "lookup_table"}

\item{refactored_fabric}{the gpkg for the corresponding VPU preceeding gpkg}
}
\value{
file path
}
\description{
Generate Lookup table for Aggregated Network
}
