% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_flowpath.R
\name{get_hw_points}
\alias{get_hw_points}
\title{Get headwater points}
\usage{
get_hw_points(fline)
}
\arguments{
\item{fline}{flowline data.frame from NHDPlus or with an ID and toID column.}
}
\description{
Returns headwater locations for line geometries
}
