% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{get_hydrofabric}
\alias{get_hydrofabric}
\title{Download Reference Fabric Data by VPU ID}
\usage{
get_hydrofabric(VPU = "01", type = "refactor", dir = NULL, overwrite = FALSE)
}
\arguments{
\item{VPU}{a VPU ID}

\item{type}{either 'refactor' (default) or 'reference'}

\item{dir}{directory path to save data to}

\item{overwrite}{should existing files be overwritten? (default = FALSE)}
}
\value{
file path
}
\description{
Download Reference Fabric Data by VPU ID
}
