% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_catchment.R
\name{get_row_col}
\alias{get_row_col}
\title{Get Row and Column}
\usage{
get_row_col(fdr, start, fac_matrix)
}
\arguments{
\item{fdr}{flow direction grid}

\item{start}{matrix (row, col)}

\item{fac_matrix}{flow accumulation matrix}
}
\description{
Get Row and Column
}
