% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hf_gpkg_from_other.R
\docType{data}
\name{gpkg}
\alias{gpkg}
\title{Convert Reference Output to HF gpkg
This is a temporary function as changes get pushed upstream}
\format{
An object of class \code{character} of length 1.
}
\usage{
gpkg
}
\arguments{
\item{gpkg}{gpkg file path}
}
\value{
file.path
}
\description{
Convert Reference Output to HF gpkg
This is a temporary function as changes get pushed upstream
}
\keyword{datasets}
