% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{hyaggregate_log}
\alias{hyaggregate_log}
\title{Logging shorthand
Log a message with given log level, and optional verbosity.}
\usage{
hyaggregate_log(level, message, verbose = TRUE)
}
\arguments{
\item{level}{log level, see logger::log_levels for more details}

\item{message}{R objects that can be converted to a character vector via the active message formatter function}

\item{verbose}{should message be emitted?}
}
\value{
log message
}
\description{
Logging shorthand
Log a message with given log level, and optional verbosity.
}
