% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{layer_exists}
\alias{layer_exists}
\title{Check if a geopackage and layer exists
This function checks if a layer exists in a geopackage}
\usage{
layer_exists(gpkg, name)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{name}{name of layer to check}
}
\value{
logical
}
\description{
Check if a geopackage and layer exists
This function checks if a layer exists in a geopackage
}
