% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hf_gpkg_from_aggregate.R
\name{make_hf_gpkg_from_aggregate}
\alias{make_hf_gpkg_from_aggregate}
\title{Convert Aggregate Output to HF gpkg
This is a temporary function as changes get pushed upstream}
\usage{
make_hf_gpkg_from_aggregate(gpkg)
}
\arguments{
\item{gpkg}{gpkg file path}
}
\value{
file.path
}
\description{
Convert Aggregate Output to HF gpkg
This is a temporary function as changes get pushed upstream
}
