% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hf_gpkg_from_other.R
\name{make_hf_gpkg_from_reference}
\alias{make_hf_gpkg_from_reference}
\title{Convert Reference Output to HF gpkg
This is a temporary function as changes get pushed upstream}
\usage{
make_hf_gpkg_from_reference(gpkg)
}
\arguments{
\item{gpkg}{gpkg file path}
}
\value{
file.path
}
\description{
Convert Reference Output to HF gpkg
This is a temporary function as changes get pushed upstream
}
