% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_outlets.R
\name{map_outlet_ids}
\alias{map_outlet_ids}
\title{Map outlets from COMID to ID for aggregate catchments}
\usage{
map_outlet_ids(source_outlets, reconciled)
}
\arguments{
\item{source_outlets}{data.frame with COMID and type columns}

\item{reconciled}{data.frame as returned by refactor workflow}
}
\description{
given reconciled flowlines and a set of source outlets,
returns a set of outlets with reconciled IDs suitable for use with 
aggregate_catchments.
}
