% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{middle_massage}
\alias{middle_massage}
\title{Re-index the interior of vector by threshold}
\usage{
middle_massage(x, index_values, threshold)
}
\arguments{
\item{x}{vector of values}

\item{index_values}{current index values}

\item{threshold}{threshold to evaluate x}
}
\value{
a vector of length(x) containing grouping indexes
}
\description{
Merges the interior values of a vector if they are less then the provided threshold.
Merging will look "up" and "down" the vector and merge into the smaller of the two.
}
