% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{network_metadata}
\alias{network_metadata}
\title{Capture Network Metadata
This function assumes  that files are names *_{VPU}.gpkg}
\usage{
network_metadata(gpkgs, flowpath_layer = "flowpaths", divide_layer = "divides")
}
\arguments{
\item{gpkgs}{a vector of file.paths to attribute}

\item{flowpath_layer}{the layer name containing flowpaths}

\item{divide_layer}{the layer name containing divides}
}
\value{
a data.frame with the file.path, file name, VPU,flowpath count, 
divides count, cumulative flowpath count and cummulative divdies count
}
\description{
Capture Network Metadata
This function assumes  that files are names *_{VPU}.gpkg
}
