% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{pack_set}
\alias{pack_set}
\title{pack set}
\usage{
pack_set(x)
}
\arguments{
\item{x}{data.frame containing "set" list column to be packed}
}
\value{
data.frame containing comma seperated character column
}
\description{
pack set
}
