% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_along_mainstems.R
\name{pinch_sides}
\alias{pinch_sides}
\title{Re-index the edges of vector by threshold
Merge the outside edges of a vector if they are less then the provides threshold.}
\usage{
pinch_sides(x, ind, thres)
}
\arguments{
\item{x}{vector of values}

\item{ind}{current index values}

\item{thres}{threshold to evaluate x}
}
\value{
a vector of length(x) containing grouping indexes
}
\description{
Re-index the edges of vector by threshold
Merge the outside edges of a vector if they are less then the provides threshold.
}
