% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_catchment.R
\name{prep_cat_fdr_fac}
\alias{prep_cat_fdr_fac}
\title{Prep catchment with FDR/FAC}
\usage{
prep_cat_fdr_fac(cat, fdr, fac)
}
\arguments{
\item{cat}{catchment (sf object)}

\item{fdr}{flow direction grid}

\item{fac}{flow accumulation grid}
}
\description{
Prep catchment with FDR/FAC
}
