% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_split_events.R
\name{prep_split_events}
\alias{prep_split_events}
\title{Prep Split Events}
\usage{
prep_split_events(pois, fline, divides, threshold = 25)
}
\arguments{
\item{pois}{a set of POIs with a poi_id, X and  (in 5070)}

\item{divides}{a set of divides geometries (EPSG:5070)}

\item{flines}{a set of flowlines geometries (EPSG:5070)}

\item{theshold}{a percentage (0-100) a POI must be upstream before splitting}
}
\value{
sf POINT object
}
\description{
Prep Split Events
}
