% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{prepare_network}
\alias{prepare_network}
\title{Prepare Hydrologic Network}
\usage{
prepare_network(network_list)
}
\arguments{
\item{network_list}{a list with flowpath and catchment data}
}
\value{
a list containing flowpath and catchment `sf` objects
}
\description{
Prepare Hydrologic Network
}
\details{
This function adds an area, length, hydrosequence, streamorder and contributing drainage area
metric to the flowpath list element of network_list.

tot_drainage_areasqkm can only be added when there are no NA areas
}
