% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{read_hydrofabric}
\alias{read_hydrofabric}
\title{Read Catchments and Flowpaths from Geopackage
Convenience function for reading two layers into a list}
\usage{
read_hydrofabric(
  gpkg = NULL,
  catchments = NULL,
  flowpaths = NULL,
  realization = "all",
  crs = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{realization}{what layers to read? Options: {"catchemnts", "flowpaths", "all"}}

\item{crs}{desired CRS, if NULL they stay as read. If all CRS layers arenot}

\item{catchment_name}{name of catchment layer. If NULL, attempts to find divides layer}

\item{flowpath_name}{name of flowpath layer. If NULL, attempts to find flowpath layer}

\item{vebose}{should message be emitted?}
}
\value{
list
}
\description{
Read Catchments and Flowpaths from Geopackage
Convenience function for reading two layers into a list
}
