% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_style.R
\name{read_qml}
\alias{read_qml}
\title{Read a QML file from the inst/qml directory}
\usage{
read_qml(name)
}
\arguments{
\item{name}{Name of QML file without extension}
}
\value{
the contents of {name}.qml.
}
\description{
Read a QML file from the inst/qml directory
}
\keyword{internal}
