% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename Geometry Column}
\usage{
rename_geometry(x, name)
}
\arguments{
\item{x}{sf data.frame}

\item{name}{character name to be used for geometry}
}
\description{
renames the geometry column of a sf object.
}
