% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{sb_id}
\alias{sb_id}
\title{Return ScienceBase ID for hydrofabric
This function checks if a layer exists in a geopackage}
\usage{
sb_id(type)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{name}{name of layer to check}
}
\value{
character
}
\description{
Return ScienceBase ID for hydrofabric
This function checks if a layer exists in a geopackage
}
