% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowline.R
\name{split_flowlines}
\alias{split_flowlines}
\title{Split Flowlines}
\usage{
split_flowlines(flines, max_length, para = 0, avoid = NA)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM
and LINESTRING sf column in "meters" projection}

\item{max_length}{maximum segment length to return}

\item{para}{numeric how many threads to use in parallel computation}

\item{avoid}{vector of ids to avoid}
}
\value{
All the flowlines with some split apart. COMIDs are returned as strings 
with a semantic part number appended. That is .1, .2, ... .10, .11, etc. are 
appended and must be treated as one would treat a semantic version. .1 is the 
most upstream and the sequence increases in the downstream direction.
}
\description{
A wrapper for split_lines that works on nhdplus attributes
}
\seealso{
The \code{\link{refactor_nhdplus}} function implements a complete
workflow using `split_flowlines()`.
}
