% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{st_rename}
\alias{st_rename}
\title{Rename simple features layer}
\usage{
st_rename(dsn, layer, new_layer)
}
\arguments{
\item{dsn}{data source name. Interpretation varies by driver: can be
a filename, a folder, a database name, or a Database Connection
(we officially test support for
\code{\link[RPostgres:Postgres]{RPostgres::Postgres()}} connections).}

\item{layer}{layer name. Varies by driver, may be a file name without
extension; for database connection, it is the name of the table. If layer
is missing, the \code{basename} of \code{dsn} is taken.}

\item{new_layer}{new layer name}
}
\value{
dsn
}
\description{
Rename simple features layer
}
