% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_catchment.R
\name{trace_upstream}
\alias{trace_upstream}
\title{Trace Upstream}
\usage{
trace_upstream(start_point, cat, fdr, fac_matrix, fdr_matrix)
}
\arguments{
\item{start_point}{row col index}

\item{cat}{catchment}

\item{fdr}{flow direction grid}

\item{fac_matrix}{flow accumulation matrix}

\item{fdr_matrix}{flow direction matrix}
}
\value{
sfc
}
\description{
Trace Upstream
}
