% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{union_linestrings_geos}
\alias{union_linestrings_geos}
\title{DEPRECATED: Fast LINESTRING union}
\usage{
union_linestrings_geos(lines, ID)
}
\arguments{
\item{lines}{lines to merge}

\item{ID}{ID to merge over}
}
\value{
an sf object
}
\description{
Wayyyy faster then either data.table, or sf based line merging
}
