% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{union_polygons_geos}
\alias{union_polygons_geos}
\title{Fast Polygon Union
This is significantly faster then sf::st_union or summarize}
\usage{
union_polygons_geos(poly, ID)
}
\arguments{
\item{poly}{sf POLYGON object}

\item{ID}{the column name over which to union geometries}
}
\value{
sf object
}
\description{
Fast Polygon Union
This is significantly faster then sf::st_union or summarize
}
