% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{unpack_set}
\alias{unpack_set}
\title{unpack set}
\usage{
unpack_set(x)
}
\arguments{
\item{x}{data.frame containing comma separated "set" column to be unpacked}
}
\value{
data.frame containing a list column
}
\description{
unpack set
}
