% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{update_network_identifiers}
\alias{update_network_identifiers}
\title{Update Network Identifiers
Given a data.frame of sf object, the id and toid values are undated based on
a provided lookup table (produced with build_new_id_table), and a vpu_topo list 
if there are cross VPU flows. In the vpu_topo is NULL or has 0 rows, no vpu correction
is applied.}
\usage{
update_network_identifiers(x, lookup, term_add = 1e+09, connections = NULL)
}
\arguments{
\item{x}{a data.frame or sf object with id and/or toid columns}

\item{lookup}{a lookup table of new ID values}

\item{vpu_topo}{a VPU lookup correction table}
}
\value{
data.frame
}
\description{
Update Network Identifiers
Given a data.frame of sf object, the id and toid values are undated based on
a provided lookup table (produced with build_new_id_table), and a vpu_topo list 
if there are cross VPU flows. In the vpu_topo is NULL or has 0 rows, no vpu correction
is applied.
}
