% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrofabric_io.R
\name{write_hydrofabric}
\alias{write_hydrofabric}
\title{Write a hydrofabric gpkg
A hydrofabric consists of a flowpath, catchment,
and topology layer written to a self contained geopackage}
\usage{
write_hydrofabric(
  network_list,
  outfile,
  catchment_name = "divides",
  flowpath_name = "flowpaths",
  edge_list_name = "edge_list",
  verbose = TRUE
)
}
\arguments{
\item{network_list}{a list containing flowpaths and catchments}

\item{outfile}{a file (gpkg) where layers should be written}

\item{catchment_name}{the layer name for divides}

\item{flowpath_name}{the layer name for flowpaths}

\item{verbose}{should messages be emitted?}
}
\value{
file path
}
\description{
Write a hydrofabric gpkg
A hydrofabric consists of a flowpath, catchment,
and topology layer written to a self contained geopackage
}
