% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nonnetwork_divides.R
\name{add_nonnetwork_divides}
\alias{add_nonnetwork_divides}
\title{Add Non Network Divides to Aggregated Fabric
The refactoring process intentionally drop catchments without a flowpath.
In cases where a seamless discritization of the landscape is needed, these 
area must be reintroduced from the reference dataset.}
\usage{
add_nonnetwork_divides(
  gpkg = NULL,
  divides = NULL,
  huc12 = NULL,
  reference_gpkg = NULL,
  verbose = TRUE
)
}
\arguments{
\item{gpkg}{a path to a gpkg}

\item{huc12}{huc12 to COMID crosswalk}

\item{reference_gpkg}{A path to the reference VPU geopackage (see get_hydrofabric(..., type = "reference"))}

\item{verbose}{Should messages be emitted?}

\item{divide}{If gpkg is NULL, then an sf data.frame, otherwise a the layer name. See details.}
}
\value{
gpkg path
}
\description{
Add Non Network Divides to Aggregated Fabric
The refactoring process intentionally drop catchments without a flowpath.
In cases where a seamless discritization of the landscape is needed, these 
area must be reintroduced from the reference dataset.
}
