% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{assign_global_identifiers}
\alias{assign_global_identifiers}
\title{Update Hydrofabric Identifiers
For a given set of hydrofabric geopackages, update the ID and toID values to 
be gloablaly unique.}
\usage{
assign_global_identifiers(
  gpkgs = NULL,
  outfiles = NULL,
  flowpath_layer = "flowpaths",
  mapped_POI_layer = "mapped_POIs",
  divide_layer = "divides",
  lookup_table_layer = "lookup_table",
  flowpath_edge_list = "flowpath_edge_list",
  overwrite = FALSE,
  update_terminals = TRUE,
  term_add = 1e+09,
  return_lookup = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gpkgs}{a vecotor of file.paths to that define the global network}

\item{outfiles}{a vector of file.paths to write to}

\item{flowpath_layer}{the layer name containing flowpaths}

\item{mapped_POI_layer}{the layer name containing flowpaths}

\item{divide_layer}{the layer name containing divides}

\item{lookup_table_layer}{the layer name containing flowpaths}

\item{overwrite}{overwrite existing files?}

\item{verbose}{emit messages}

\item{catchment_network_layer}{the layer name containing flowpaths}
}
\value{
a data.frame
}
\description{
Update Hydrofabric Identifiers
For a given set of hydrofabric geopackages, update the ID and toID values to 
be gloablaly unique.
}
