% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_hydrofabrics.R
\name{build_new_id_table}
\alias{build_new_id_table}
\title{Build a new ID lookup table}
\usage{
build_new_id_table(x, flowpath_layer = "flowpaths", divide_layer = "divides")
}
\arguments{
\item{flowpath_layer}{the layer name containing flowpaths}

\item{divide_layer}{the layer name containing divides}

\item{gpkgs}{a row of network metadata built with `network_metadata`}
}
\value{
a data.frame with the file.path, file name, VPU,flowpath count, 
divides count, cumulative flowpath count and cumulative divides count
}
\description{
Build a new ID lookup table
}
