% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{clean_geometry}
\alias{clean_geometry}
\title{Clean Catchment Geometry}
\usage{
clean_geometry(catchments, ID = "ID", keep = 0.9)
}
\arguments{
\item{catchments}{catchments geometries to fix}

\item{ID}{name of uniquely identifying column}

\item{keep}{proportion of points to retain in geometry simplification 
(0-1; default 0.05). See \code{\link[rmapshaper]{ms_simplify}}. 
If NULL, then no simplification will be executed.}
}
\value{
sf object
}
\description{
Fixes geometry issues present in catchments that originate in the 
CatchmentSP layers, or from the reconcile_catchments hyRefactor preocess. 
These include, but are not limited to disjoint polygon fragments, artifacts 
from the 30m DEM used to generate the catchments, and non-valid geometry topolgies.
A goal of this functions is also to provide means to reduce the data column 
of the catchments by offering a topology preserving simplification 
through \code{\link[rmapshaper]{ms_simplify}}. 
Generally a "keep" parameter of .9 seems appropriate for the resolution of 
the data but can be modified in function
}
