% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyaggregate_utils.R
\name{flush_prefix}
\alias{flush_prefix}
\title{Flush existing ID prefixes
Given a data object and column, remove a prefix and adjoining "-"}
\usage{
flush_prefix(input, col)
}
\arguments{
\item{input}{input data object}

\item{col}{column to remove prefix from}
}
\value{
data object with updated column
}
\description{
Flush existing ID prefixes
Given a data object and column, remove a prefix and adjoining "-"
}
