% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchment_geometry.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename a sf geometry column}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{sf object}

\item{name}{desried name of geometry column}
}
\value{

}
\description{
Rename an `sf` geometry column without prior knowledge of current name. See \href{https://gis.stackexchange.com/questions/386584/sf-geometry-column-naming-differences-r}{here}
}
\examples{
library(sf)
nc = st_read(system.file("shape/nc.shp", package="sf"))
names(nc) # shows "geom"
nc = rename_geometry(nc, "gyom")
names(nc) # shows "gyom"
}
