% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_catchment.R
\name{split_catchment_divide}
\alias{split_catchment_divide}
\title{Split Catchment Divides}
\usage{
split_catchment_divide(
  catchment,
  fline,
  fdr,
  fac,
  lr = FALSE,
  min_area_m = 800,
  snap_distance_m = 100,
  simplify_tolerance_m = 40,
  vector_crs = NULL
)
}
\arguments{
\item{catchment}{sf data.frame with one catchment divide}

\item{fline}{sf data.frame with one or more flowline segments in
upstream downstream order.}

\item{fdr}{SpatRaster a flow direction SpatRaster that fully covers the catchment}

\item{fac}{SpatRaster a flow accumulation SpatRaster that fuller covers the catchment}

\item{lr}{boolean should catchments be split along the left/right bank?}

\item{min_area_m}{minimum area in m^2 to filter out slivers (caution, use with care!!)}

\item{snap_distance_m}{distance in meters to snap SpatRaster generated geometry to polygon geometry}

\item{simplify_tolerance_m}{dTolerance in meters for simplification of grid-cell based polygons}

\item{vector_crs}{any object compatible with sf::st_crs. Used for vector-based calculations in case that 
fdr projection is not suitable (e.g. lon/lat) -- must result in units of meters.}
}
\value{
Split catchment divides as an sfc geometry.
}
\description{
A catchment-divide splitting algorithm that works with a D8
flow direction grid and the output of nhdplus_refactor. See Vignette
for examples.
}
