% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_subsetting.R
\name{subset_network}
\alias{subset_network}
\title{Subset the upstream portion of a network}
\usage{
subset_network(
  gpkg,
  origin,
  flowpath_edgelist = "flowpath_edge_list",
  flowpath_name = "flowpaths",
  catchment_name = "divides",
  mainstem = FALSE,
  attribute_layers = NULL,
  export_gpkg = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gpkg}{path to a hydrofabric}

\item{origin}{the ID to begin navigation}

\item{flowpath_edgelist}{layer name of flowpath edgelist in gpkg}

\item{flowpath_name}{layer name of flowpaths in gpkg}

\item{catchment_name}{layer name of cathcments in gpkg}

\item{mainstem}{should only the mainstem flowpath be retruned (default = FALSE)}

\item{export_gpkg}{a path to write the data to. If NULL a list is returned}
}
\description{
Subset the upstream portion of a network
}
